backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-multipart==0.0.17
Pillow==11.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
aiofiles==24.1.0
